requirements_fastapi.txt
# FastAPI Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Async HTTP client (optional, for future enhancements)
httpx==0.25.1
aiofiles==23.2.1

# Existing dependencies
flask==3.0.0
google-cloud-vision==3.4.5
google-generativeai==0.3.1
python-dotenv==1.0.0
PyMuPDF>=1.23.0

# Session management
itsdangerous==2.1.2

# Database & Authentication
sqlalchemy==2.0.23
alembic==1.12.1

# FastAPI Users & Authentication
fastapi-users[sqlalchemy]==12.1.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Templates for Admin Panel
jinja2==3.1.2
